Added MASM and GAS Assembler tests
