+++ lua/compiler.lua
@@ -1,53 +1,16 @@
 require("lua.util")
 
-function GetExecutableNames(name)
-	return {name, "a"}
-end
+function CreateCTestFile(name, src)
+	if not src then src = "int main(void) {\n\treturn 0;\n}\n" end
 
-function GetExecutableExtensions()
-	return {".com", ".exe", ".out"}
-end
-
-function GetCompilerOutputExtensions()
-	return {".a", ".com", ".dll", ".exe", ".lib", ".o", ".obj", ".out"}
-end
-
-function CheckAndReturnCommonExecutable(baseName)
-	local names = GetExecutableNames(baseName)
-	local extensions = GetExecutableExtensions()
-
-	for _, name in ipairs(names) do
-		for _, extension in ipairs(extensions) do
-			local fileName = name .. extension
-			local file = io.open(fileName)
-
-			if file then
-				file:close()
-				return fileName
-			end
-		end
-	end
-end
-
-function CheckAndRemoveCommonArtifacts(baseName)
-	local names = GetExecutableNames(baseName)
-	local extensions = GetCompilerOutputExtensions()
-	local r = 0
-
-	for _, name in ipairs(names) do
-		for _, extension in ipairs(extensions) do
-			local fileName = name .. extension
-			local file = io.open(fileName)
-
-			if file then
-				file:close()
-				os.remove(fileName)
-				r = r + 1
-			end
-		end
+	local file = io.open(name, "w")
+	if file then
+		file:write(src)
+		file:close()
+		return true
 	end
 
-	return r
+	return false
 end
 
 function CheckCustomCompiler(cc, tmpName)
